### 2020-10-08 a09cf8c
[JOSS] Corrected invalid DOI
