+++ JOSS-paper/paper.bib
@@ -90,7 +90,7 @@ number = "2",
 pages = "312 - 349",
 year = "2018",
 issn = "1063-5203",
-doi = "https://doi.org/10.1016/j.acha.2016.05.002",
+doi = "10.1016/j.acha.2016.05.002",
 url = "http://www.sciencedirect.com/science/article/pii/S1063520316300185",
 author = "James Bremer",
 }
