### 2020-10-08 #2729 @fruzsinaagocs
Hi @kthyng  :wave:
The invalid DOI whedon discovered has been corrected.
I think the following people could potentially be reviewers:
- lgarrison
- danfortunato

From the volunteer list:
- harpolea
- danehkar
- fhorrobin
- rgmyr
- bhajay
- solanpaa
- ziotom78
- corentin-dev
- jakryd
