+++ README.rst
@@ -132,29 +132,6 @@ and then include the relevant header files in your C++ code:
     #include "solver.hpp"
     #include "system.hpp"
 
-Installation troubleshooting
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-1. Eigen import errors:
-    .. code:: bash
-
-       pyoscode/_pyoscode.hpp:6:10: fatal error: Eigen/Dense: No such file or directory
-        #include <Eigen/Dense>
-                  ^~~~~~~~~~~~~
-
-    Try explicitly including the location of your Eigen library via the
-    ``CPLUS_INCLUDE_PATH`` environment variable, for example:
-
-    .. code:: bash
-
-       CPLUS_INCLUDE_PATH=/usr/include/eigen3 python setup.py install --user
-       # or 
-       CPLUS_INCLUDE_PATH=/usr/include/eigen3 pip install pyoscode
-
-    where  ``/usr/include/eigen3`` should be replaced with your system-specific
-    eigen location.
-
-
 
 Quick start
 -----------
