### 2020-10-23 8ece94d
Updated docstring for python interface
