### 2020-10-28 #2729 @melissawm
Hello, @corentin-dev, @jakryd would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Let me know if you have need any more information. Thanks!
