### 2020-11-03 #2729 @melissawm
Hello @jgoldfar! Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Let me know if you need any more information. Thanks!
