### 2020-11-09 #2729 @melissawm
Hello, @dlfivefifty ! Thank you for agreeing to review this software paper for JOSS! I'll assign you as reviewer, and whedon (our bot) will close this issue and open a new [REVIEW] issue. We'll meet there for the remainder of the process.
