### 2020-11-09 #2830 @whedon
[REVIEW]: (py)oscode: fast solutions of oscillatory ODEs
**Submitting author:** @fruzsinaagocs (<a href="http://orcid.org/0000-0002-1763-5884">Fruzsina Julia Agocs</a>)
**Repository:** <a href="https://github.com/fruzsinaagocs/oscode" target ="_blank">https://github.com/fruzsinaagocs/oscode</a>
**Version:** v1.0.3
