### 2020-11-12 f04564b
Added wand to Dockerfile
