### 2020-11-12 c401921
Trying to add hide_code
