+++ Dockerfile
@@ -3,8 +3,9 @@ FROM ubuntu:18.04
 RUN apt-get update && apt-get install -y \
     git python3-pip libmagickwand-dev
 RUN pip3 install --no-cache --upgrade pip && \
+    pip3 install nbconvert==5.6.1 && \
     pip3 install --no-cache notebook && \
-    pip3 install nbconvert==5.6.1 scipy numpy matplotlib ipywidgets jupyter_contrib_nbextensions \
+    pip3 install scipy numpy matplotlib ipywidgets jupyter_contrib_nbextensions \
     hide_code cite2c RISE wand && \
     jupyter contrib nbextension install --user && \
     jupyter nbextension enable hide_code --user --py && \
