### 2020-11-12 3c7f317
Trying to add hide_code 3
