### 2020-11-12 8c3aab5
Trying to add hide_code 4
