+++ Dockerfile
@@ -8,7 +8,8 @@ RUN pip3 install --no-cache --upgrade pip && \
     pip3 install scipy numpy matplotlib ipywidgets jupyter_contrib_nbextensions \
     hide_code cite2c RISE wand && \
     jupyter nbextension install --py --user hide_code &&\
-    jupyter nbextension enable --py --sys-prefix hide_code &&\
+    jupyter nbextension enable --py --user hide_code &&\
+    jupyter serverextension enable --py --user hide_code &&\
     jupyter contrib nbextension install --user && \
     python3 -m cite2c.install
 
