+++ .travis.yml
@@ -1,8 +1,7 @@
 os:
     - linux
     - osx
-    osx_image:
-    - xcode12.2
+osx_image:  xcode12.2
 language: python
 python: 
     - "2.7"
