### 2020-12-02 62a112e
Migrated to travis-ci.com, reproducing #5
