+++ .travis.yml
@@ -1,5 +1,4 @@
 os:
-    - linux
     - osx
 osx_image:  xcode12.2
 language: python
