### 2020-12-02 cbfa5fa
Removed python lines from mac travis config
