+++ .travis.yml
@@ -1,12 +1,6 @@
 os:
     - osx
 osx_image:  xcode12.2
-language: python
-python: 
-    - "2.7"
-    - "3.4"
-    - "3.5"
-    - "3.6"
 install: 
     - pip install -r requirements.txt
     - pip install pyoscode
