+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["numpy", "setuptools>=40.8.0", "wheel"]
+build-backend = "setuptools.build_meta"
