+++ setup.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, with_statement, print_function, division
 from setuptools import setup, Extension, find_packages
 import os
-import numpy.distutils.misc_util
+import numpy as np
 
 def readme(short=False):
     with open("README.rst") as f:
@@ -13,14 +13,14 @@ def readme(short=False):
 pyoscode_module = Extension(
     name="_pyoscode",
     sources=["pyoscode/_pyoscode.cpp"],
-    include_dirs=['include','pyoscode',numpy.distutils.misc_util.get_numpy_include_dirs()],
+    include_dirs=['include','pyoscode',np.get_include()],
     depends=["pyoscode/_python.hpp", "pyoscode/_pyoscode.hpp"],
     extra_compile_args=['-std=c++11','-Wall']
     )
 
 setup(
     name="pyoscode",
-    version="1.0.1",
+    version="1.0.2",
     description=readme(short=True),
     long_description=readme(),
     url="https://github.com/fruzsinaagocs/oscode",
@@ -37,7 +37,7 @@ setup(
     license="oscode",
     ext_modules=[pyoscode_module],
     headers=["pyoscode/_python.hpp", "pyoscode/_pyoscode.hpp"],
-    include_dirs=numpy.distutils.misc_util.get_numpy_include_dirs(),
+    include_dirs=[np.get_include()],
     keywords="PPS, cosmic inflation, cosmology, oscillatory, ODE",
     classifiers=[
                 'Intended Audience :: Developers',
