+++ README.rst
@@ -256,9 +256,11 @@ devs of `exhale <https://pypi.org/project/exhale/>`__ for making the beautiful C
 Changelog
 ---------
 
-- 1.0.1: current version
+- 1.0.2: current version
+    - Fixed getting correct numpy include directories
+- 1.0.1:
     - Added `pyproject.toml` to handle build dependencies (numpy)
-- 1.0.0: first release
+- 1.0.0:
     - Dense output
     - Arrays for frequency and damping term need not be evenly spaced
     - Automatic C++ documentation on readthedocs
