+++ .travis.yml
@@ -6,7 +6,7 @@ python:
     - "3.7"
 install: 
     - pip install -r requirements.txt
-    - pip install pyoscode==1.0.1
+    - pip install pyoscode==1.0.2
 script:
     - coverage run -m pytest tests/
 branches:
