### 2020-12-03 ede380e
Added pytest to reqs
