+++ requirements.txt
@@ -1,4 +1,5 @@
 numpy
 scipy
 matplotlib
+pytest
 coverage
