### 2020-12-03 3b227e4
Adding linux back in for travis
