### 2020-12-04 #2830 @dlfivefifty
I finished the review, everything looks fine. @fruzsinaagocs it's outside the scope of this review, but it would be interesting to compare performance with ApproxFun.jl, which also efficiently solves ODEs. E.g. look at the Airy example in the README:

https://github.com/JuliaApproximation/ApproxFun.jl

ApproxFun.jl doesn't do anything special for highly oscillatory problems, but, it is `O(n)` complexity for `n` degrees of freedom, and is pretty fast (around a second for a million unknowns).

More interesting would be to do something special a la ultraspherical spectral method for the oscillatory case.  I did something for ODEs ages ago:

https://link.springer.com/chapter/10.1007%2F978-3-642-15337-2_24

Work by Keller gives some ideas for combining with the ultraspherical spectral method used in ApproxFun.jl. If this is of interest please let me know.
