### 2020-12-11 #2830 @jakryd
Thanks for re-inviting me. I finished the review, and I think that the package meets the JOSS requirements and format.

Congratulations @fruzsinaagocs!
