+++ JOSS-paper/paper.bib
@@ -151,7 +151,7 @@ pages={79–101}
 }
 
 @misc{rkwkb-handley,
-    title={The Runge-Kutta-Wentzel-Kramers-Brillouin Method},
+    title={{The Runge-Kutta-Wentzel-Kramers-Brillouin Method}},
     author={Handley, W. J. and Lasenby, A. N. and Hobson, M. P.},
     year={2016},
     eprint={1612.02288},
