### 2020-12-11 #2830 @fruzsinaagocs
Thank you @melissawm, I fixed a capitalisation error in a reference and tweaked a sentence (really minor change to clarify that I'm referring to the expansion of the universe). With these changes, the paper's good to go!
