### 2020-12-14 #2830 @melissawm
Hello, @fruzsinaagocs ! A few small issues remain:

- in [Deaño, Huybrechs, Iserles (2017)] the publisher is `Society for Industrial; Applied Mathematics` (there is an error rendering the `&` character, I believe)
- There is a citation by [Hu Li, B et al] that doesn't seem to correspond to the actual article. I think this is the right one (and also includes a DOI that is missing from the reference list): https://www.emerald.com/insight/content/doi/10.1108/03321640910999897/full/html Please confirm if this is the case.
- Some of the references contain full journal names (e.g. SIAM Journal on Numerical Analysis) and some contain abbreviated versions of the journal name (e.g. Phys. Rev. D). It would be better to have the full name for all journals as the practice of abbreviations varies between disciplines and some readers may not be familiar with that.
- A citation to [SciPy](https://scipy.org/citing.html) is missing in the summary.

Let me know when these are sorted so we can do a final check. Thanks!
