### 2020-12-14 9170c70
Fixing remaining bibliography issues
