+++ JOSS-paper/paper.bib
@@ -1,7 +1,31 @@
+@article{scipy,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+            Kern, Robert and Larson, Eric and Carey, C J and
+            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+            Harris, Charles R. and Archibald, Anne M. and
+            Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+            {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2},
+}
+
 @article{oscode,
   title = {Efficient method for solving highly oscillatory ordinary differential equations with applications to physical systems},
   author = {Agocs, F. J. and Handley, W. J. and Lasenby, A. N. and Hobson, M. P.},
-  journal = {Phys. Rev. Research},
+  journal = {Physical Review Research},
   volume = {2},
   issue = {1},
   pages = {013030},
@@ -16,7 +40,7 @@
 @article{beyond-rkwkb,
   title = {{Beyond the Runge-Kutta-Wentzel-Kramers-Brillouin method}},
   author = {Bamber, Jamie and Handley, Will},
-  journal = {Phys. Rev. D},
+  journal = {Physical Review D},
   volume = {101},
   issue = {4},
   pages = {043517},
@@ -31,7 +55,7 @@
 @article{pps-curved,
   title = {Primordial power spectra for curved inflating universes},
   author = {Handley, Will},
-  journal = {Phys. Rev. D},
+  journal = {Physical Review D},
   volume = {100},
   issue = {12},
   pages = {123517},
@@ -46,7 +70,7 @@
 @article{qic,
   title = {Quantum initial conditions for inflation and canonical invariance},
   author = {Agocs, F. J. and Hergt, L. T. and Handley, W. J. and Lasenby, A. N. and Hobson, M. P.},
-  journal = {Phys. Rev. D},
+  journal = {Physical Review D},
   volume = {102},
   issue = {2},
   pages = {023507},
@@ -95,23 +119,24 @@ url = "http://www.sciencedirect.com/science/article/pii/S1063520316300185",
 author = "James Bremer",
 }
 
-@article{hu-et-al-circuits,
+@article{condon-et-al-circuits,
 issn = {0332-1649},
+doi = {10.1108/03321640910999897}
 journal = {Compel},
 pages = {1607--1618},
 volume = {28},
-publisher = {Boole Press Limited,},
+publisher = {Boole Press Limited},
 number = {6},
 year = {2009},
+url = {https://doi.org/10.1108/03321640910999897},
 title = {On numerical methods for highly oscillatory problems in circuit simulation},
-address = {[Dublin] :},
[…]
