+++ JOSS-paper/paper.bib
@@ -121,7 +121,7 @@ author = "James Bremer",
 
 @article{condon-et-al-circuits,
 issn = {0332-1649},
-doi = {10.1108/03321640910999897}
+doi = {10.1108/03321640910999897},
 journal = {Compel},
 pages = {1607--1618},
 volume = {28},
