### 2020-12-14 397afd9
Trying to keep the 'and' in journal name
