+++ JOSS-paper/paper.bib
@@ -136,7 +136,7 @@ author = {Condon, Marissa and Deaño, Alfredo and Iserles, Arieh and Maczyński,
 @book{deano-integrals,
 author = {Deaño, Alfredo and Huybrechs, Daan and Iserles, Arieh},
 title = {Computing Highly Oscillatory Integrals},
-publisher = {{Society for Industrial and Applied Mathematics}},
+publisher = {{Society for Industrial {and} Applied Mathematics}},
 year = {2017},
 doi = {10.1137/1.9781611975123},
 address = {Philadelphia, PA},
