### 2020-12-14 #2830 @fruzsinaagocs
Thanks! I think this is all done now.

> * in [Deaño, Huybrechs, Iserles (2017)] the publisher is `Society for Industrial; Applied Mathematics` (there is an error rendering the `&` character, I believe)

Interestingly, bibTeX didn't like the word 'and' here.

> * There is a citation by [Hu Li, B et al] that doesn't seem to correspond to the actual article. I think this is the right one (and also includes a DOI that is missing from the reference list): https://www.emerald.com/insight/content/doi/10.1108/03321640910999897/full/html Please confirm if this is the case.

Well spotted! I corrected this now, and it should have a doi.

> * Some of the references contain full journal names (e.g. SIAM Journal on Numerical Analysis) and some contain abbreviated versions of the journal name (e.g. Phys. Rev. D). It would be better to have the full name for all journals as the practice of abbreviations varies between disciplines and some readers may not be familiar with that.

All corrected to full journal names.

> * A citation to [SciPy](https://scipy.org/citing.html) is missing in the summary.

Added.
