+++ JOSS-paper/paper.bib
@@ -130,7 +130,7 @@ number = {6},
 year = {2009},
 url = {https://doi.org/10.1108/03321640910999897},
 title = {On numerical methods for highly oscillatory problems in circuit simulation},
-author = {Condon, Marissa and Deaño, Alfredo and Iserles, Arieh and Maczyński, Kornel},
+author = {Condon, M. and Deaño, A. and Iserles, A. and Maczyński, K.},
 }
 
 @book{deano-integrals,
