### 2020-12-15 #2830 @melissawm
Hi @fruzsinaagocs one last thing, I think according to the DOI link the citation for [Condon et. al. 2009] should be

Condon, M., Deaño, A., Iserles, A., Maczyński, K. and Xu, T. (2009), "On numerical methods for highly oscillatory problems in circuit simulation", COMPEL - The international journal for computation and mathematics in electrical and electronic engineering, Vol. 28 No. 6, pp.

(note the full journal name)

As far as I can tell this is the last step! Thanks again.
