+++ JOSS-paper/paper.bib
@@ -122,7 +122,8 @@ author = "James Bremer",
 @article{condon-et-al-circuits,
 issn = {0332-1649},
 doi = {10.1108/03321640910999897},
-journal = {Compel},
+journal = {{COMPEL - The international journal for computation {and} mathematics
+    in electrical {and} electronic engineering}},
 pages = {1607--1618},
 volume = {28},
 publisher = {Boole Press Limited},
