### 2020-12-15 #2830 @kthyng
Hi @fruzsinaagocs! I'm taking over to finish publication your work. Can you verify the version of the software you want associated with it? I see the newest version on the repo is v1.0.3.
