### 2020-12-15 #2830 @fruzsinaagocs
@kthyng Yes, the version number is v1.0.3!
