### 2020-12-15 #2830 @kthyng
@fruzsinaagocs Ok great, I'll update that.

Regarding your paper, just a few notes:
* I fixed some inline references in a [PR](https://github.com/fruzsinaagocs/oscode/pull/6), so please merge if you agree with them
* I can't quite tell how to fix the last sentence before the acknowledgments. If the reference is to be inline (without parentheses) then take the square brackets off. If it is meant to be parenthetically referenced, then keep the square brackets. You can have multiple references in one set of brackets, separated by a semi-colon. [More info](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
