### 2020-12-15 af17fee
Fixed some more inline citations
