+++ JOSS-paper/paper.bib
@@ -158,7 +158,7 @@ pages={79–101}
 } 
 
 @misc{bremer-code,
-  author = {Bremer},
+  author = {Bremer, J.},
   title = {{Phase functions: Fortran 90 code for solving highly oscillatory
   ordinary differential equations in O(1) time}},
   year = {2018},
