### 2020-12-15 bd2f366
Fixing Bremer citation again
