+++ JOSS-paper/paper.bib
@@ -107,16 +107,16 @@ pages={437–483}
 }
 
 @article{bremer,
-title = "On the numerical solution of second order ordinary differential equations in the high-frequency regime",
-journal = "Applied and Computational Harmonic Analysis",
-volume = "44",
-number = "2",
-pages = "312 - 349",
-year = "2018",
-issn = "1063-5203",
-doi = "10.1016/j.acha.2016.05.002",
-url = "http://www.sciencedirect.com/science/article/pii/S1063520316300185",
-author = "James Bremer",
+title = {{On the numerical solution of second order ordinary differential equations in the high-frequency regime}},
+journal = {Applied {and} Computational Harmonic Analysis},
+volume = {44},
+number = {2},
+pages = {312 - 349},
+year = {2018},
+issn = {1063-5203},
+doi = {10.1016/j.acha.2016.05.002},
+url = {http://www.sciencedirect.com/science/article/pii/S1063520316300185},
+author = {Bremer, J.},
 }
 
 @article{condon-et-al-circuits,
