### 2020-12-15 #2830 @fruzsinaagocs
Thanks @kthyng, I've merged your PR and corrected another inline citation + the last citation.
